+++ src/main/java/ru/sberbank/homework/dragonblog/service/UserService.java
@@ -6,9 +6,9 @@ import ru.sberbank.homework.dragonblog.util.NotFoundException;
 import java.util.List;
 
 public interface UserService {
-    User get(long id)  throws NotFoundException;
+    User get(int id)  throws NotFoundException;
 
-    boolean delete(long id)  throws NotFoundException;
+    boolean delete(int id)  throws NotFoundException;
 
     User update(User user)  throws NotFoundException;
